JIP-195 REFACT: 예약하기를 bookId 에서 bookInfoId 로 변경
